packets c., spell pos and angle

+++ Packets/S2C/0xD8-ATTACH_MINIMAP_ICON.js
@@ -3,7 +3,7 @@ var BasePacket = require('../BasePacket');
 
 module.exports = class extends BasePacket {//S2C.
 	struct = {
-		UnitNetID: 'uint8',
+		UnitNetID: 'uint32',
 		ChangeIcon: 'uint8',
 		IconCategory: ['char', 64],
 		ChangeBorder: 'uint8',
